requirements.txt
absl-py==1.0.0
alembic==1.7.7
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.0
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
cssselect==1.1.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fake-useragent==0.1.11
fastapi==0.75.1
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h11==0.13.0
h5py==3.6.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
lxml==4.8.0
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parse==1.19.0
passlib==1.7.4
pluggy==1.0.0
protobuf==3.20.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pyee==8.2.2
pyparsing==3.0.8
pyppeteer==1.0.2
pyquery==1.4.3
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
w3lib==1.22.0
watchgod==0.8.2
websockets==10.2
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
